renamed to better fit descriptions, refactored
