+++ app/src/main/java/bg/zahov/app/MyApplication.kt
@@ -12,7 +12,7 @@ import bg.zahov.app.data.provider.SettingsProviderImpl
 import bg.zahov.app.data.provider.UserProviderImpl
 import bg.zahov.app.data.provider.WorkoutProviderImpl
 import bg.zahov.app.data.provider.WorkoutStateManager
-import bg.zahov.app.data.provider.WorkoutTopBarProvider
+import bg.zahov.app.data.provider.WorkoutActionHandler
 import bg.zahov.app.ui.exercise.topbar.ExerciseTopBarManager
 import com.google.firebase.Firebase
 import com.google.firebase.initialize
@@ -55,7 +55,7 @@ object Inject {
         ExerciseTopBarManager.getInstance()
     }
     val workoutTopAppHandler by lazy {
-        WorkoutTopBarProvider.getInstance()
+        WorkoutActionHandler.getInstance()
     }
 }
