+++ app/src/main/java/bg/zahov/app/ui/topbar/TopBar.kt
@@ -20,7 +20,7 @@ import bg.zahov.app.Settings
 import bg.zahov.app.StartWorkout
 import bg.zahov.app.Workout
 import bg.zahov.app.ui.exercise.topbar.TopBarExercise
-import bg.zahov.app.ui.workout.TopBarWorkout
+import bg.zahov.app.ui.workout.topbar.TopBarWorkout
 import bg.zahov.fitness.app.R
 import kotlin.reflect.KClass
